Transform Graph model in a service and transform graphcontroller in dijkstrasController + dijkstra service"
