+++ app/services/graph_service.rb
@@ -1,0 +1,15 @@
+class GraphService
+  attr_reader :vertices
+
+  def initialize
+    @vertices = {}
+  end
+
+  def add_vertex(name, edges)
+    @vertices[name] = edges
+  end
+
+  def to_s
+    @vertices.inspect
+  end
+end
